+++ src/lib/converter/converter.ts
@@ -229,7 +229,21 @@ export class Converter extends ChildableComponent<
     constructor(owner: Application) {
         super(owner);
 
-        this.addUnknownSymbolResolver((ref) => {
+        const userConfiguredSymbolResolver: ExternalSymbolResolver = (
+            ref,
+            refl,
+            _part,
+            symbolId,
+        ) => {
+            if (symbolId) {
+                return userConfiguredSymbolResolver(
+                    symbolId.toDeclarationReference(),
+                    refl,
+                    undefined,
+                    undefined,
+                );
+            }
+
             // Require global links, matching local ones will likely hide mistakes where the
             // user meant to link to a local type.
             if (ref.resolutionStart !== "global" || !ref.symbolReference) {
@@ -256,7 +270,9 @@ export class Converter extends ChildableComponent<
             if (typeof modLinks["*"] === "string") {
                 return modLinks["*"];
             }
-        });
+        };
+
+        this.addUnknownSymbolResolver(userConfiguredSymbolResolver);
     }
 
     /**