+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+### Bug Fixes
+
+-   `externalSymbolLinkMappings` now uses the TypeScript reported link target if available, #2725.
+
 ## v0.26.8 (2024-10-04)
 
 ### Features